Restructuring of `CallExpression` and `MemberCallExpression`
This PR restructures the way call expressions and member call expressions work. It introduces a new `callee` field, to allow modelling of calls that were previously not possible (see #689) and also reworks the way, the `base` property works.

This IS API breaking and will probably the first PR that starts the v5 development.

Fixes #689
Fixes #289